[REFACTOR] MetroStopComparator 리팩토링
- 파일 로드하는 파트는 별도의 MetroStopFileService 로 분리하였음
- 기존 지하철역과 비교한 후 upsert하는 기능을 MetroStopSyncService 가 가져갔음
