+++ src/main/java/org/hanihome/hanihomebe/metro/application/MetroStopSyncScheduler.java
@@ -4,8 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
-import org.hanihome.hanihomebe.metro.application.service.MetroStopComparator;
-import org.hanihome.hanihomebe.metro.application.service.MetroStopService;
+import org.hanihome.hanihomebe.metro.application.service.MetroStopSyncService;
 import org.hanihome.hanihomebe.metro.infra.MetroStopApiClient;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -16,7 +15,7 @@ import java.io.IOException;
 @RequiredArgsConstructor
 @Component
 public class MetroStopSyncScheduler {
-    private final MetroStopComparator metroStopComparator;
+    private final MetroStopSyncService metroStopSyncService;
     private final MetroStopApiClient metroStopApiClient;
 
     @Scheduled(cron = "0 0 0 * * *", zone = "UTC")    // 매일 00시
@@ -24,7 +23,7 @@ public class MetroStopSyncScheduler {
         log.info("지하철역 동기화 스케줄링 시작");
         try {
             metroStopApiClient.fetchAndStoreMetroStops();
-            metroStopComparator.checkAndUpdateStops();
+            metroStopSyncService.checkAndUpdateStops();
         } catch (IllegalStateException | IOException e) {
             throw new CustomException(ServiceCode.METRO_STOP_SYNCHRONIZE_FAILED, e);
         }
